// neuralbets/src/types/Livegame.ts
export interface live_game {
    id: string;
    sport_name: string;
    sport_title: string;
    home_team: string;
    home_team_id: string;
    away_team: string;
    away_team_id: string;
    market: string;
    bookmaker: string;
    home_team_price: number;
    away_team_price: number;
    start_time: string;
}